+++ v1/frontend/package.json
@@ -9,6 +9,7 @@
     "start": "npm run dev",
     "test:unit": "jest --config test/unit/jest.conf.js --coverage",
     "test:unit:debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --config test/unit/jest.conf.js --no-cache --runInBand",
+    "test:unit:watch": "jest --config test/unit/jest.conf.js --coverage --watch",
     "test": "npm run stylelint && npm run lint && npm run test:unit",
     "lint": "eslint --ext .js,.vue src test/unit",
     "stylelint": "stylelint .scss, .vue, .css",
@@ -47,6 +48,7 @@
     "eslint-plugin-vue": "^4.0.0",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.4",
+    "flush-promises": "^1.0.2",
     "friendly-errors-webpack-plugin": "^1.6.1",
     "html-webpack-plugin": "^2.30.1",
     "jest": "^22.0.4",
