+++ v1/frontend/src/assets/scss/_elements.scss
@@ -1,7 +1,7 @@
 body {
   box-sizing: border-box;
   margin: 0;
-  background-color: rgb(238, 197, 72);
+  background: $almost-black;
   font-family: $font-base;
 }
