+++ v1/frontend/src/assets/scss/_variables.scss
@@ -8,7 +8,12 @@ $font-size-s: 16px;
 
 $primary-color: rgba(#EEC548, 1);
 $secondary-color: rgba(#505050, 1);
+$primary-gradient:
+  linear-gradient(to bottom, #EAA91C 0%, #EEC548 25%, #EEC548 39%, #EAA91C 100%);
 $black: #000;
+$almost-black: #222;
 $white: #FFF;
 $soft-white: rgba($white, 0.9);
 
+$header-height: 155px;
+$mid-height: 76px + 65px;
