+++ v1/frontend/src/components/movies/MovieDetail.vue
@@ -22,7 +22,6 @@
 
 <script>
 import { mapActions, mapGetters } from 'vuex';
-
 export default {
   data() {
     return {
