+++ v1/frontend/src/components/movies/__tests__/Movies.spec.js
@@ -1,7 +1,7 @@
-import { createLocalVue, shallowMount } from '@vue/test-utils';
+import { createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
 import VueRouter from 'vue-router';
-import Movies from '../Movies';
+// import Movies from '../Movies';
 import Header from '../../generic/Header';
 
 const localVue = createLocalVue();
@@ -10,27 +10,21 @@ localVue.use(VueRouter);
 localVue.component('v-header', Header);
 
 describe('Movies.vue', () => {
-  let storeOptions;
-  let store;
-
-  beforeEach(() => {
-    storeOptions = {
-      getters: {
-        getMovies: jest.fn(),
-        loggedIn: jest.fn(),
-      },
-    };
-    store = new Vuex.Store(storeOptions);
+  // let storeOptions;
+  // let store;
+  //
+  // beforeEach(() => {
+  //   storeOptions = {
+  //     getters: {
+  //       getMovies: jest.fn(),
+  //       loggedIn: jest.fn(),
+  //     },
+  //   };
+  //   store = new Vuex.Store(storeOptions);
+  // });
+  test('shows listed movie-items', () => {
   });
-
-  test('render image based on url', () => {
-    const $getMovies = [];
-    const wrapper = shallowMount(Movies, {
-      mocks: { $getMovies },
-      localVue,
-      store,
-    });
-    expect(wrapper.find('ul').text()).toContain('');
+  test('shows search bar when logged in', () => {
   });
 });
