+++ v1/frontend/src/store/__tests__/getters.spec.js
@@ -1,0 +1,10 @@
+import getters from '../getters';
+
+describe('getters', () => {
+  test('token returns the same value form state', () => {
+    const token = 'eenhelelangestring';
+    const state = { token };
+    const result = getters.token(state);
+    expect(result).toEqual(token);
+  });
+});
